Move setup_volume to volume_ops
