+++ common/mixin.py
@@ -9,13 +9,12 @@ from .ops.support_ops.io_ops import IoOps
 from .ops.support_ops.machine_ops import MachineOps
 from .ops.gluster_ops.peer_ops import PeerOps
 from .ops.gluster_ops.volume_ops import VolumeOps
-from .ops.gluster_ops.volume_libs import VolumeLibs
 from .ops.gluster_ops.gluster_ops import GlusterOps
 from .ops.gluster_ops.brick_ops import BrickOps
 from .ops.gluster_ops.profile_ops import ProfileOps
 
 
-class RedantMixin(GlusterOps, VolumeLibs, VolumeOps, BrickOps, PeerOps,
+class RedantMixin(GlusterOps, VolumeOps, BrickOps, PeerOps,
                   IoOps, MachineOps, ProfileOps, Rexe, Logger):
     """
     A mixin class for redant project to encompass all ops, support
